Nested property handling (#72)
* adding functionality for nested property handling to provide more flexibility.
